pyproject.toml
[project]
name = "seichijunrei"
version = "0.1.0"
description = "AI-powered travel assistant for anime pilgrims using Gemini LLM"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Seichijunrei Team", email = "team@seichijunrei.com"}
]
keywords = ["anime", "travel", "ai", "gemini", "seichijunrei", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Dependencies
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Async/HTTP
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",

    # Google APIs
    "googlemaps>=4.10.0",
    "google-cloud-storage>=2.10.0",

    # Observability
    "structlog>=23.0.0",
    "rich>=13.0.0",

    # Google Cloud & ADK
    "google-genai>=1.0.0",
    "google-cloud-aiplatform[adk,agent_engines]>=1.111",
    "google-adk>=1.23.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",

    # Code Quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["application", "clients", "config", "domain", "infrastructure", "services", "utils"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    ".uv_python/*",
    "*/.uv_python/*",
    ".uv_cache/*",
    "*/.uv_cache/*",
    ".uv-cache/*",
    "*/.uv-cache/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
